lesson13/vaia-l13-context-engineering/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
tiktoken==0.5.2
nltk==3.8.1
scikit-learn==1.4.0
numpy==1.26.3
aiohttp==3.9.1
prometheus-client==0.19.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
